add change as reviewed
